requirements.txt
twitchio==2.9.1
python-dotenv==1.0.0
asyncio-throttle==1.0.2
